Update server code to new operator interface
Redesign serde code

+++ presto-main/src/main/java/com/facebook/presto/nblock/uncompressed/UncompressedBlock.java
@@ -62,9 +62,9 @@ public class UncompressedBlock
         return rawRange;
     }
 
-    public long getCount()
+    public int getCount()
     {
-        return range.getEnd() - range.getStart() + 1;
+        return (int) (range.getEnd() - range.getStart() + 1);
     }
 
     @Override
